Changed NOTIFYICON_VERSION to NOTIFYICON_VERSION_XP and added NOTIFYICON_VERSION_VISTA, added hBalloonIcon to NOTIFYICONDATA for custom balloon icon, changed return type from GetVersion to uint32 (in win docs it its DWORD and DWORD is uint32), added Set_hBalloonIcon to NOTIFYICONDATA.

+++ kernel32.go
@@ -141,7 +141,6 @@ func init() {
 	mulDiv = MustGetProcAddress(libkernel32, "MulDiv")
 	setLastError = MustGetProcAddress(libkernel32, "SetLastError")
 	systemTimeToFileTime = MustGetProcAddress(libkernel32, "SystemTimeToFileTime")
-
 }
 
 func CloseHandle(hObject HANDLE) bool {
@@ -264,12 +263,12 @@ func GetThreadUILanguage() LANGID {
 	return LANGID(ret)
 }
 
-func GetVersion() int64 {
+func GetVersion() uint32 {
 	ret, _, _ := syscall.Syscall(getVersion, 0,
 		0,
 		0,
 		0)
-	return int64(ret)
+	return uint32(ret)
 }
 
 func GlobalAlloc(uFlags uint32, dwBytes uintptr) HGLOBAL {
